sub admin details edit api updated

+++ src/resolvers/adminResolver/adminResolver.ts
@@ -17,7 +17,8 @@ export const adminResolver: Resolvers = {
 
     // create super admin and sub admin 
     createAdmin: async (parent, { input, image }, { req }, info) => {
-
+ 
+       await verifySuperAdmin(req)
       await validateInput(validators.AdminCreateValidator, req);
       try {
 
@@ -97,7 +98,7 @@ export const adminResolver: Resolvers = {
 
     suspendAdmin: async (parent, { input }, { req }, info) => {
 
-
+      await verifySuperAdmin(req)
       try {
 
         await adminModel.findByIdAndUpdate({ _id: input?.id }, {
@@ -443,6 +444,7 @@ export const adminResolver: Resolvers = {
 
     deleteAdminAccount: async (parent, { input }, { req }, info) => {
 
+      await verifySuperAdmin(req)
       try {
 
         await adminModel.findByIdAndDelete({ _id: input?.id })
@@ -469,7 +471,7 @@ export const adminResolver: Resolvers = {
 
     editAdminAccount: async (parent, { input,image }, { req }, info) => {
 
-
+      // await verifySuperAdmin(req)
       await validateInput(validators.AdminEditValidator, req);
 
 
@@ -494,8 +496,13 @@ export const adminResolver: Resolvers = {
 
          const isEmailExists=await adminModel.findOne({email:email})
 
-         if(isEmailExists && isEmailExists._id!==input.id){
+         console.log("input id",input.id.toString())
 
+         console.log("email exit",isEmailExists?._id.toString())
+
+         if(isEmailExists && isEmailExists._id.toString()!==input?.id.toString()){
+
+           console.log("exit")
 
           throw new GraphQLError('Admin with this email already exists', {
             extensions: {
@@ -572,7 +579,7 @@ export const adminResolver: Resolvers = {
 
     cretaeDeliveryAgentConfig: async (parent, { input }, { req }, info) => {
 
-
+    await verifyAdmin(req)
       try {
 
         const limit = input.limit || 0
@@ -601,7 +608,8 @@ export const adminResolver: Resolvers = {
 
 
     updateDeliveryAgentConfig: async (parent, { input }, { req }, info) => {
-      // await verifyAdmin(req)
+     
+      await verifyAdmin(req)
 
       try {
         const { deliveryLimit, returnLimit, _id } = input
@@ -693,6 +701,7 @@ export const adminResolver: Resolvers = {
 
     getAllAdminData: async (parent, { input }, { req }, info) => {
 
+       await verifySuperAdmin(req)
       try {
 
         const page: number = input?.page || 0;
@@ -726,6 +735,8 @@ export const adminResolver: Resolvers = {
 
     getOneAdminDetails: async (parent, { input }, { req }, info) => {
 
+
+         await verifySuperAdmin(req)  
 
       try {
